added security for deleting and saveOrUpdateUserForm  for users

+++ src/main/java/com/roman_musijowski/pgs_lessons/services/repoServicesImp/UserServiceRepoImp.java
@@ -3,7 +3,9 @@ package com.roman_musijowski.pgs_lessons.services.repoServicesImp;
 import com.roman_musijowski.pgs_lessons.commands.UserForm;
 import com.roman_musijowski.pgs_lessons.converters.UserFormToUser;
 import com.roman_musijowski.pgs_lessons.models.User;
+import com.roman_musijowski.pgs_lessons.models.security.Role;
 import com.roman_musijowski.pgs_lessons.repositories.UserRepository;
+import com.roman_musijowski.pgs_lessons.services.RoleSevice;
 import com.roman_musijowski.pgs_lessons.services.UserService;
 import com.roman_musijowski.pgs_lessons.services.security.EncryptionService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,12 +21,14 @@ public class UserServiceRepoImp implements UserService {
     private UserRepository userRepository;
     private UserFormToUser userFormToUser;
     private EncryptionService encryptionService;
+    private RoleSevice roleSevice;
 
     @Autowired
-    public UserServiceRepoImp(UserRepository userRepository, UserFormToUser userFormToUser, EncryptionService encryptionService) {
+    public UserServiceRepoImp(UserRepository userRepository,  RoleSevice roleSevice, UserFormToUser userFormToUser, EncryptionService encryptionService) {
         this.userRepository = userRepository;
         this.userFormToUser = userFormToUser;
         this.encryptionService = encryptionService;
+        this.roleSevice = roleSevice;
     }
 
     @Override
@@ -43,8 +47,10 @@ public class UserServiceRepoImp implements UserService {
     @Override
     public User saveOrUpdate(User object) {
         if(object.getPassword() != null){
+            System.out.println("Save or update User");
             object.setEncryptedPassword(encryptionService.encryptString(object.getPassword()));
         }
+        System.out.println("User updated");
 
         return userRepository.save(object);
     }
@@ -56,6 +62,11 @@ public class UserServiceRepoImp implements UserService {
 
     @Override
     public void delete(Long id) {
+
+        if (getById(id).getUserName() == "admin@gmail.com"){
+            System.out.println("You can't delete admin");
+            return;
+        }
         userRepository.deleteById(id);
     }
 
@@ -66,18 +77,30 @@ public class UserServiceRepoImp implements UserService {
 
     @Override
     public User saveOrUpdateUserForm(UserForm userForm) {
+        List<Role> roles = (List<Role>) roleSevice.listAll();
 
         User newUser = userFormToUser.convert(userForm);
 
-        if (newUser.getId() != null) {
+
+        if (newUser.getId() == null) {
+
+            roles.forEach(role -> {
+                if (role.getRole() == "STUDENT") {
+                    System.out.println("add new role student");
+                    newUser.addRole(role);
+                }
+            });
+        }else {
+            System.out.println("Existing user ! ");
             User existingUser = getById(newUser.getId());
 
             newUser.setEnabled(existingUser.getEnabled());
             newUser.setRoles(existingUser.getRoles());
             newUser.setLessons(existingUser.getLessons());
 
-            System.out.println("Roles - "+existingUser.getRoles());
-            System.out.println("Lessons - "+existingUser.getLessons());
+//            System.out.println("Roles - "+existingUser.getRoles());
+//            System.out.println("Lessons - "+existingUser.getLessons());
+            System.out.println("Added info to existing user!!");
         }
 
         return saveOrUpdate(newUser);
